removes unneeded component
