+++ modules/components/src/AdvancedSqonBuilder/sqonPieces/common.js
